requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
Django==5.2.9
django-crispy-forms==2.5
django-daraja==1.3.0
idna==3.11
pillow==12.0.0
pycparser==2.23
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
